+++ internal/core/custom_resources/resources/cloud_security_tables.go
@@ -28,8 +28,8 @@ import (
 	"go.uber.org/zap"
 
 	cloudsecglue "github.com/panther-labs/panther/internal/compliance/awsglue"
-	"github.com/panther-labs/panther/internal/core/pantherdb"
 	"github.com/panther-labs/panther/internal/log_analysis/awsglue"
+	"github.com/panther-labs/panther/internal/log_analysis/pantherdb"
 	"github.com/panther-labs/panther/pkg/awsutils"
 )
