+++ internal/log_analysis/athenaviews/athenaviews_test.go
@@ -24,10 +24,10 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/panther-labs/panther/internal/core/pantherdb"
 	"github.com/panther-labs/panther/internal/log_analysis/awsglue"
 	"github.com/panther-labs/panther/internal/log_analysis/log_processor/parsers"
 	"github.com/panther-labs/panther/internal/log_analysis/log_processor/parsers/awslogs"
+	"github.com/panther-labs/panther/internal/log_analysis/pantherdb"
 )
 
 type table1Event struct {
